Método para salvar e-mail do professor

+++ src/main/java/com/example/service/TipoUsuarioService.java
@@ -22,7 +22,8 @@ public class TipoUsuarioService {
     }
     
   
-    public TipoUsuario buscarPorId(Integer id) {
+    public TipoUsuario buscarPorId(Long id) {
         return tipoUsuarioRepository.findById(id).orElse(null);
     }
+
 }
